+++ web/apps/cast/src/types/file/index.ts
@@ -1,4 +1,4 @@
-import type { Metadata } from "@/media/types/file";
+import type { Metadata } from "@/media/file-metadata";
 import type {
     EncryptedMagicMetadata,
     MagicMetadataCore,
