+++ web/apps/photos/src/services/upload/uploadService.ts
@@ -3,9 +3,9 @@ import { basename } from "@/base/file";
 import log from "@/base/log";
 import { CustomErrorMessage } from "@/base/types/ipc";
 import { hasFileHash } from "@/media/file";
+import type { Metadata } from "@/media/file-metadata";
 import { FileType, type FileTypeInfo } from "@/media/file-type";
 import { encodeLivePhoto } from "@/media/live-photo";
-import type { Metadata } from "@/media/types/file";
 import { cmpNewLib, extractExif, wipNewLib } from "@/new/photos/services/exif";
 import * as ffmpeg from "@/new/photos/services/ffmpeg";
 import type { UploadItem } from "@/new/photos/services/upload/types";
